requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
